+++ tests/testthat/test-block.R
@@ -1,7 +1,7 @@
 context("test-block.R")
 
-test_that("block yaml works", {
-  expect_error(block_yaml(license = "foo"), msg = "license not legal")
+test_that("vw_block_config yaml works", {
+  expect_error(vw_block_config(license = "foo"), msg = "license not legal")
   expect_identical(
     vw_block_config(),
     "license: mit\nheight: 500\nscrolling: yes\nborder: yes\n"
@@ -14,7 +14,7 @@ test_that("block index works", {
     readLines("../block/index.html") %>%
     paste(collapse = "\n")
 
-  index <- block_index(
+  index <- vw_block_index(
     version = list(vega = "1.2.3", vega_lite = "4.5.6", vega_embed = "7.8.9"),
     embed = list(defaultStyle = TRUE, renderer = "canvas")
   )