+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * overhauls the API (#15, with @AliciaSchep)
 
+  * changes `block_build_directory()` -> `vw_block_build_directory()`
+  * changes `block_index()` -> `vw_block_index()`
+  * changes `block_create()` -> `vw_create_block()`
   * changes `block_config()` -> `vw_block_config()`  
   * changes `autosize()` -> `vw_autosize()`
   * changes `as_json()` -> `vw_as_json()`